relaxed version constraints
